refactor(notes): centralize private sync schema
