+++ backlog/docs/sqlite-private-owner-inventory.md
@@ -68,7 +68,7 @@ Classifications have these meanings:
 | C47 | tldw_chatbook/TTS/profile_repository | restore version/qualification/publication source | tts.profile_restore_stage | read_only_uri | exact admitted restore source | Migrated via `connect_private_sqlite`. Every restore-stage connection is repository-owned under the exclusive lease. A close failure retains the live handle and exclusive lease, leaves the repository unavailable, and is retried only by later cleanup. |
 | C48 | tldw_chatbook/TTS/profile_schema | open_exact_current_profile_store | tts.profile_store_descriptor | read_only_uri | immutable descriptor-bound shared-startup proof | Migrated via `connect_private_sqlite_descriptor`. Shared startup retains the canonical parent/file descriptors, validates an immutable exact-v4 logical image, and requires the query-only path connection to serialize to the same image before writes are enabled; the returned live handle owns the proof descriptors for its lifetime. |
 | C49 | tldw_chatbook/DB/Subscriptions_DB | SubscriptionsDB._get_connection | db.subscriptions.agent_read | read_only_uri | external agent Watchlists read | Migrated via `connect_private_sqlite`. Opens only an existing Watchlists database through a read-only URI, preserves the source file mode owned by the mutable application database, and cannot create or migrate the database or write the main database file, schema, or rows. A WAL reader may create or update SQLite-managed `-wal`/`-shm` sidecars; suppressing that with `immutable=1` could ignore committed, uncheckpointed WAL frames. |
-| C50 | tldw_chatbook/Notes/note_import_receipts | NoteImportReceiptRepository._connect | notes.sync_state | private_file | device-private import receipts and future lasting-sync state | Migrated via `connect_private_sqlite`. The profile-local ledger stores only opaque identifiers, private digests, bounded lifecycle state, and reconciliation metadata; it is excluded from portable export and centralized backup. |
+| C50 | tldw_chatbook/Notes/notes_sync_state_schema | notes_sync_state_transaction | notes.sync_state | private_file | device-private import receipts and future lasting-sync state | Migrated via `connect_private_sqlite`. The profile-local ledger stores only opaque identifiers, private digests, bounded lifecycle state, and reconciliation metadata; it is excluded from portable export and centralized backup. |
 
 ## SQLite backup and restore inventory
